+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 
 ## Unreleased
+### Added
+- Data Link checksum calculation using "+?"
 ### Changed
 - Finalized Data Link - bytes are reversed from "!" version - `utils/byte_reverse.py` can fix existing Data Link byte sequences in text files
 - Finalized Fusion Loader - removed "!!" with no changes